Entities created, login fixed

+++ src/main/java/edu/eci/cosw/pancomido/service/UserService.java
@@ -12,7 +12,7 @@ public interface UserService
 {
     List<User> getUsers();
 
-    User getUser( Long id );
+    User getUser( Integer id );
 
     User createUser( User user );
